data-extraction/requirements.txt
streamlit
pdfplumber
numpy
faiss-cpu
sentence-transformers
transformers
rouge-score
